requirements.txt
celery==5.4.0
cryptography==42.0.7
Django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-filter==24.2
bcrypt==4.1.3
psycopg2==2.9.9
PyJWT==2.8.0
redis==5.0.4
